# Project/Deliverables/PA/docs/openapi/a9_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'SportsVerse Web API'
  description: 'Web Resources Specification (A9) for SportsVerse'
servers:
- url: http://lbaw2223.lbaw.fe.up.pt/
  description: Production server
...
paths:
...
  /category:
    get:
      operationId: R202
      summary: 'R202: Product categories'
      description: 'Shows all the product categories available. Access: PUB.'
      tags:
        - 'M02: Products and Categories'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                example:
                  - id: 1
                    name: Protein
                  - id: 2
                    name: Whey
        '500':
          description: 'Error showing Home Page'
...
  /api/category:
    post:
      operationId: R306
      summary: 'R306: Add category Action'
      description: 'Adds a new category to the website. Access: ADM'
      tags:
        - 'M03: Management Area'
 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
              example:
                - name = Protein

      responses:
        '302':
          description: 'Redirect after processing the new category.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successfully created category. Redirect to admin page.'
                  value: '/profile'
                302Error:
                  description: 'Failed to create category. Redirect to the add category page.'
                  value: '/category/add'
...